+++ src/app/components/menus/collection-side/collection-side-menu.component.ts
@@ -9,7 +9,7 @@ import { CollectionPagePathPipe } from '@pipes/collection-page-path.pipe';
 import { CollectionPagePositionQueryparamPipe } from '@pipes/collection-page-position-queryparam.pipe';
 import { CollectionTableOfContentsService } from '@services/collection-toc.service';
 import { ScrollService } from '@services/scroll.service';
-import { addOrRemoveValueInArray, isBrowser } from '@utility-functions';
+import { addOrRemoveValueInArray, enableFrontMatterPageOrTextViewType, isBrowser } from '@utility-functions';
 
 
 @Component({
@@ -42,11 +42,6 @@ export class CollectionSideMenuComponent implements OnInit, OnChanges, OnDestroy
     private scrollService: ScrollService,
     private tocService: CollectionTableOfContentsService
   ) {
-    this.enableCover = config.collections?.frontMatterPages?.cover ?? false;
-    this.enableTitle = config.collections?.frontMatterPages?.title ?? false;
-    this.enableForeword = config.collections?.frontMatterPages?.foreword ?? false;
-    this.enableIntroduction = config.collections?.frontMatterPages?.introduction ?? false;
-
     this.sortSelectOptions = {
       header: $localize`:@@CollectionSideMenu.SortOptions.SelectSorting:Välj sortering för innehållsförteckningen`,
       cssClass: 'custom-select-alert'
@@ -64,7 +59,10 @@ export class CollectionSideMenuComponent implements OnInit, OnChanges, OnDestroy
           changes.collectionID.previousValue !== changes.collectionID.currentValue
         ) {
           // Collection changed, the new menu will be loaded in the subscription
-          // in ngOnInit(), so no need to do anything here.
+          // in ngOnInit(). Update collection frontmatter pages if collectionID set.
+          if (this.collectionID) {
+            this.updateFrontmatterPages();
+          }
           break;
         } else if (
           (
@@ -93,6 +91,8 @@ export class CollectionSideMenuComponent implements OnInit, OnChanges, OnDestroy
   }
 
   ngOnInit() {
+    this.updateFrontmatterPages();
+
     // Subscribe to BehaviorSubject emitting the current TOC.
     // The received TOC is already properly ordered.
     this.tocSubscr = this.tocService.getCurrentCollectionToc().subscribe(
@@ -120,6 +120,13 @@ export class CollectionSideMenuComponent implements OnInit, OnChanges, OnDestroy
 
   ngOnDestroy() {
     this.tocSubscr?.unsubscribe();
+  }
+
+  private updateFrontmatterPages() {
+    this.enableCover = enableFrontMatterPageOrTextViewType('cover', this.collectionID, config);
+    this.enableTitle = enableFrontMatterPageOrTextViewType('title', this.collectionID, config);
+    this.enableForeword = enableFrontMatterPageOrTextViewType('foreword', this.collectionID, config);
+    this.enableIntroduction = enableFrontMatterPageOrTextViewType('introduction', this.collectionID, config);
   }
 
   private updateHighlightedMenuItem(scrollTimeout: number = 600) {